Change local lock system to make it robust against shutdown
Previously when the device running the script has been powered down,
the lock file was not deleted and so the backup wrongly stayed locked.

Now using the proper tempfs in /var/run for writing the lock files
(also changed to proper PID files as intended in this FS) files will
be automatically deleted at shutdown and cannot interfere with a
shutdown anymore.