Simplify script
Now the check checking the process name has been removed.
This could be problematic on some systems and is not really reliable, so
we just verify that the process is running for now. It would be really
unlucky if the same PID is given to another process in such a special
case anyway.