Implement additional notification details like size/time of backup
It turned out the way of parsing the borg create output	is hard	to do
in POSIX and you need to have at least one tempfile.
A way better way is to just query borg after the backup	for the	details
and calculate the time e.g. by yourself. Yes, even though we have to do	that
it is easier than parsing borg.	In bash/zsh it would maybe be possible
to catch and parse the output, but well…

Also moved more information into the title in order to have space for all
taht additional information we put into the description.